making newer InertiaRelief with easier to use EqualityConstrainedHomotopyProblem class the default
