+++ examples/inertia_relief/CMakeLists.txt
@@ -11,12 +11,6 @@ if (SERAC_ENABLE_CONTINUATION AND STRUMPACK_DIR)
                         DEPENDS_ON  serac_physics continuation_solver
                         )
     
-    blt_add_executable( NAME        inertia_relief_example2
-                        SOURCES     inertia_relief_example2.cpp
-                        OUTPUT_DIR ${EXAMPLE_OUTPUT_DIRECTORY}
-                        DEPENDS_ON  serac_physics continuation_solver
-                        )
-
     serac_add_example_test(NAME            inertia_relief
                            COMMAND         inertia_relief_example
                            NUM_MPI_TASKS   4)
@@ -24,7 +18,6 @@ if (SERAC_ENABLE_CONTINUATION AND STRUMPACK_DIR)
     install(
         FILES
             inertia_relief_example.cpp
-	    inertia_relief_example2.cpp
         DESTINATION
             examples/serac/inertia_relief
     )
